backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
httpx==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
celery[redis]==5.3.6
redis==5.0.1
gitpython==3.1.41
radon==6.0.1
reportlab==4.1.0
python-docx==1.1.0
python-multipart==0.0.9
PyJWT==2.8.0
cryptography==42.0.2
aiofiles==23.2.1
requests==2.31.0
